Finished Weather API

+++ PlanFort/PlanFort/Models/APIModels/OpenWeatherAPIModel/WeatherResponseModel.cs
@@ -24,24 +24,24 @@ namespace PlanFort.Models.APIModels.OpenWeatherAPIModel
 
     public class Coord
     {
-        public float lon { get; set; }
-        public float lat { get; set; }
+        public double lon { get; set; }
+        public double lat { get; set; }
     }
 
     public class Main
     {
-        public float temp { get; set; }
-        public float feels_like { get; set; }
-        public float temp_min { get; set; }
-        public float temp_max { get; set; }
+        public double temp { get; set; }
+        public double feels_like { get; set; }
+        public double temp_min { get; set; }
+        public double temp_max { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
     }
 
     public class Wind
     {
-        public float speed { get; set; }
-        public int deg { get; set; }
+        public double speed { get; set; }
+        public double deg { get; set; }
     }
 
     public class Clouds
